finance-app/auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
pydantic-settings
python-jose[cryptography]
httpx
email-validator
mangum